media-source/SourceBuffer-buffered_removed.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>Reading SourceBuffer#buffered after the sourceBuffer object is removed</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var FILE = 'test.webm';
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vorbis,vp8"';

function GET(url, processBody) {
    var xhr = new XMLHttpRequest();
    xhr.open('GET', url, true);
    xhr.responseType = 'arraybuffer';
    xhr.send();
    xhr.onload = function(e) {
        if (xhr.status != 200) {
            alert("Unexpected status code " + xhr.status + " for " + url);
            return false;
        }
        processBody(new Uint8Array(xhr.response));
    };
}
async_test(function(t) {
    GET(FILE, function(data) {
        var mediaSource = new MediaSource();
        var num_updateend = 0;
        mediaSource.addEventListener('sourceopen', function(e){
            var sourceBuffer = mediaSource.addSourceBuffer(mime);
            sourceBuffer.addEventListener('updateend', function(e) {
                if(!num_updateend) {
                    t.step(function() {
                        assert_throws({name: 'InvalidStateError'},
                                      function() {
                                          mediaSource.removeSourceBuffer(sourceBuffer);
                                          sourceBuffer.buffered;
                                      },
                                      'After removing the sourceBuffer object, read the buffered attribute');
                    });
                    t.done();
                }
                num_updateend++;
            });
            sourceBuffer.appendBuffer(data);
        });
        video.src = window.URL.createObjectURL(mediaSource);
    });
},
'SourceBuffer#buffered : If this object has been removed from the sourceBuffers attribute of the parent media source, then INVALID_STATE_ERR exception should be thrown');
</script>
</body>
</html>
